requirements.txt
alive_progress==3.2.0
matplotlib==3.10.3
numpy==2.3.1
openai==1.96.1
opencv_python==4.11.0.86
Pillow==11.3.0
pupil_apriltags==1.0.4.post11
pyaudio==0.2.14
pyfiglet==1.0.2
pygame==2.6.1
xarm==0.0.4
xarm_python_sdk==1.15.0
